New unrand: toga "Victory"
This is an early unrand with a variable bonus based on kills:

As the player kills threatening monsters, this robe gradually gains AC and
Slay, capping out at +10 each.

Beware -- Quaffing potions or reading scrolls resets these bonuses to zero!

(This still needs art: It's just using the robe of clouds tile for now.)