Xom-themed Crypt end
Undead Xom worshipers have built a shrine deep in the Crypt (or maybe they just took over a temple to Yred/Kiku?)  Some of the followers are wielding chaos weapons, a few dancing draining weapons are wielding themselves, and neqoxecs and cacodemons will provide plenty of opportunity for hilarious mutations.  The temple is led by a mummified Xom priest in the back.