Remove obsolete Pakellas code
There's no reason for all this to still exist, since Pakellas doesn't
pretend to be anywhere near functional if a player does somehow transfer a
Pakellas-worshipping save to a modern version, and any future
reimplementation of the god would need to be a redesign from scratch.
Just keep around the minimum amount of code to allow transferring without
crashing or completely breaking the game, and don't punish players for
abandoning Pakellas.