Fix some Dowan/Duvessa speech handling when dueled
If it's even possible to duel one of Dowan/Duvessa, fixes the surviving
twin acting as if the dueled twin had been banished. If the player abandons
Okawaru mid-duel with one of the twins, the other will still act like the
dueled twin is dead on their return, but this is probably impossible to do
in the first place anyway.