Fix various Step from Time pathfinding interactions (acrobat)
Don't let monsters path towards the player during Step from Time if marked
or if (0,0) was visible from the location where the ability was used, and
don't let monsters re-target the player while they're at (0,0) if they fail
a stealth check.

It's entirely feasible that this makes Step from Time work completely
differently to how it used to, but since nobody really understood how it
worked in the first place that's probably fine.