fix: correctly sanitize mouse hover in memorize menu
With menu_arrow_control=false, hover can still be set by the mouse, but
`set_hover` does not apply the sanitizing `min` check (because it
doesn't change it at all). This commit adds a corresponding check for
this case in `update_menu`, which should be called whenever
`items.size()` may have changed.