Increase sticky flame damage by 25% vs monsters
Sticky Flame has taken some hits in recent versions, not least due to
attacks of opportunities weakening kiting. Now that it's not in the
Fire Elementalist start, it can afford to be a bit stronger.

This doesn't affect the displayed damage, which only shows the initial
impact. TODO: clean up and unify this.

This does not affect sticky flame damage on players.