fix: Remove an unecessary subvault (Goratrix)
The vault grunt_ashenzari_visionary_trees is one of two fallback vaults
placed when information about Lair branches isn't available. However we
don't need two distinct vaults for this case, as fallback never happens
in an actual game. Remove the trees vault, leaving
grunt_ashenzari_visionary_lava as the fallback, and add allow_dup to the
latter so it can be used twice.