fix: explicit handling for macro raw input mode
This seems to have only been working accidentally, by passing everything
to the server. But this relied on the server sending back hover info,
which seems to be broken by recent changes [I am not sure this was
intentional]. Anyways, it's better to just manually implement this so
that the client knows it is in manual input mode and can act
accordingly, rather than having everything work by side effect.