fix: explicit default foreground for to_colour_string
When the crawl binary renders colour strings, it uses lightgray as the
default foreground color. However, this is not explicitly part of the
string, so the webtiles color parsing code was not inserting a lightgray
span, leading to artifacts. This commit ensures that this default is
present in all autogenerated colour strings; it may or may not have some
side effects I haven't noticed. (N.b. trying to change this on the
webtiles side breaks a *lot* of formatting, unfortunately.)