feat: Serial library vaults
Add a new serial vault type that places a large number of vaults that use the
wall_studio tile set. Most vaults are small and relatively simple, but a few
large vaults that place thematic monsters and/or book loot are included. The
vaults can place in D, Snake, Elf, and Depths at 0.5% chance.