fix: correct webtiles hidpi implementation
This commit reverts 3925e268bea408.

To get hidpi to work in a browser canvas object, the internal dimensions
need to be twice that of the external dimensions (essentially). The
normal way of implementing this is to multiply the internal dimensions
by devicePixelRatio, and scale the canvas accordingly. We used to do
this, but 3925e268bea40 changed things so that the external dimensions
were halved. This was, according to the commit message, in order to
prevent scaling artifacts on firefox. (Where the salient fact is that
window.devicePixelRatio on some browsers, these days including
chrome/firefox but not safari, also changes with browser zoom.) However,
a consequence is that tile rendering is halved on hidpi devices, exactly
not what should happen. This change also made for a lot of messy code
that had to interact directly with devicePixelRatio.

The half-size tile rendering wasn't immediately noticeable because the
dungeon is auto-scaled, and the monster list matches its scale to the
dungeon. But, icons in menus and popups instantiate the problem, and
scaling them to the dungeon size is not ideal. The action panel also had
some workarounds for this that I didn't fully understand at the time. (I
only noticed this myself after many years of working on this stuff
because I switched from a regular-density monitor to a 4k monitor of
similar size, even though I've used retina devices for a long time.
Hidpi devices have been broken for so long that I won't be surprised if
we get requests to readd this behavior as an option or something, menus
may look odd now to people who have only ever played on hidpi.)

In my testing so far, I can't really detect any rendering differences
between the approach after this commit and the approach in
3925e268bea408. However, I'm not exactly sure of what the original
issues were, and this commit should be tested on a non-hidpi device
before merge.