feat: add an option to reduce redraws for animations
The reduce_beam_redraw option now decides whether to draw whole beam at
once instead of animating every tile seperately with screen redraws in
between. To achieve this the bolt::draw function is modified to return
after setting the tile to the bolt but before trying to redraw. The
redraw is then done only once outside the loop in which bolt::draw
function is called to draw the whole beam at once.

[committer's note: previously, the only choices for animations were to
have the full animation (at a customized rate), or to show nothing at
all. This provides a useful intermediate option that will be
particularly valuable on webtiles. Note that a following commit will
rename this option to `reduce_animations`.]