Add reduce_bream_redraw option to draw beams for animation without many redraws
The reduce_beam_redraw option now decides whether to draw whole beam at once
instead of animating every tile seperately with screen redraws in between. To achieve
this the bolt::draw function is modified to return after setting the tile to the bolt but
before trying to redraw. The redraw is then done only once outside the loop in which
bolt::draw function is called to draw the whole beam at once.