fix: remove empty spans in newgame buttons
Because these are using a css flex layout with equal sizes, empty spans
generated from a color string have a weird effect on these buttons.
Since 29c30a7bfa80, the strings used for these buttons are guaranteed to
have an empty span. Possibly the approach in that commit should be
revisited, and I'm unclear on why the button spans are formatted this
way, but for now, fix the local problem in the newgame buttons by
removing empty spans.