fix: update hover on local tiles click
If you hover with a mouse, then use arrows to move the hover away, then
click without moving the mouse at all, the currently hovered item (not
under the mouse) gets activated. This commit updates the hover on
mousedown, so that when the click is processed (on mouse up), the item
under the mouse is what does get clicked.