feat: readable keycodes for some non-ascii keys
This is a fairly inelegant approach that gets the job done for now; it
lets "F1" (etc) for function keys, "NPx" for numpad key x, as well as
"Down", "PgDn", etc for navigation keys, work as keycodes. These can
appear in contexts where a \{} keycode is valid, as well as directly in
the [] brackets in a `bindkey` rc line.