Make Refrigeration and Ramparts check AC and do more damage
One of the trickier things to teach new players is that AC works
against elemental attacks. Fireball, bolt of cold, etc - AC will help!
It's a simple, consistent mechanic that has big implications across
play.

This becomes much trickier, and more confusing, when an arbitrary
subset of spells do *not* check AC. What, ultimately, is the difference
between a flash freeze, a freezing cloud, and a Refrigeration? Why should
the latter ignore AC while the former two don't? Can't all that thick
armour insulate you from the cold?

If a spell doesn't have a lot else going on in terms of mechanical
distinctiveness, we can have it ignore AC as its one big Thing. That's how
Freeze works, and that's fine. But adding it into spells that have many
other mechanical distinctions is both unnecessary and confusing. Let's not.

Freezing Ramparts and Ozocubu's Refrigeration both now check AC and do more
damage (+6 for Framparts, +10 for Refrigeration), based on a vague average
of AC in relevant parts of the game.