feat: generalize `F` to fire any targeted item
This gives a quicker way of accessing e.g. the targeter for scroll of
fear than quivering it. Behavior on launchers and ammo remains the same
as before, and you can still use `*` to toss any item.

There's a lot of special cases to deal with here in terms of
preconditions, so it may be that there are more cases where this menu
should be prevented altogether.

Also, fix a bug where untargeted firing of disabled items would show the
targeter prompt; now it just messages right away. (This is pretty niche,
previously you would have to do something like quiver a potion of magic
to see this.)