Experimental: redesign Ozocubu's Refrigeration
In DCSS, spells are most exciting when they're context sensitive.
You cast Fireball or Freezing Cloud when the terrain and enemy
positions suit it. It's fine to have strong spells, but it's not
very fun to have one spell which is so universally good that you can
repeatedly cast it to clear the great majority of fights. Ozo's Fridge
fell in this category.

Let's experiment with making it a bit more situation-dependent. This
commit increases its damage by ~10% and removes the -Potion gimmick, but
reduces damage against creatures that are adjacent to others (including
the player). The current formula reduces damage by 25% if there's one
creature adjacent and by 50% if there are two or more, but this should
be considered very much a first draft.

Concerns I've seen about this design:
1. It'll cause players to lure enemies so you can fight one-on-one. This
   is generally wise, but I don't think a screen-wide AOE particularly
   pushes players to do this, even if it's worse against clumps of enemies.
2. It'll be hard for players to tell what's happening. I've done some
   work on the targeter to try to show how much enemies are affected, and
   probably some work could be done on printed messages if necessary.