fix: limit vaults builder tag effects to vaults
This is a followup to c37658a0fea4; the effect here isn't really a false
positive, in that it can impact seed stability on local games when the
binary isn't restarted. The situation is that Vaults layouts, in order
to place vaults used elsewhere, typically need to ensure they are
non-opaque to the builder. The means by which it does this (tags like
transparent and passable), then become part of the vault's state, and
persist to any later reuse. 2170ac5abdc97a already limited the effect of
`passable` to uses when placing a vault in Vaults, but the v_rooms.lua
code (for reasons that are not entirely clear to me) used the
`transparent` tag to a similar effect. Changing this to `passable` has
the same effect and prevents the seed stability issues (at least, in my
testing).

It would be nice if there were a better way of resetting vault tags to
the des cache state without restarting the binary?

This commit also fixes what seems to have been a logic error in the
interaction between passable and opaque, from 2170ac5abdc97a, where the
passable tag would not result in MMT_PASSABLE outside vaults, but would
also prevent MMT_OPAQUE where appropriate. (This was probably creating
a whole bunch of unnecessary vetoes for affected vaults..)