feat: mode cycling for use item menus
This lets you cycle between wield/wear/puton, and read/quaff (depending
on starting mode). It's a pretty rough and ready implementation, but
recent menu changes made it possible to implement this way. The cycle
keys are the usual ones (arrow keys, !), but it's currently not hinted
-- e.g. the feature is secret for now. Ideally, a similar change would
happen for the remove menus before this becomes fully baked, but that's
a bit more work, since those are regular `InvMenu`s.