Allow harm to stack
Harm is a pretty silly and niche effect. Currently, multiple sources
of Harm don't stack. It's rare that you'll have multiple usable sources
of Harm, but on the occasion that you do, it seems like it'd be fun and
distinctive to allow the effect to intensify somewhat.

The formula used is as follows:
- Damage dealt: 130% at one level of Harm, 145% at two levels.
- Damage taken: 120% at one level of Harm, 130% at two levels.

This is somewhat conservative. One level of Harm already pushes the game
a bit toward 'rocket tag', in which one side or another of a fight dies
instantly, and we don't want the effect to totally turn the game into a
mess. But it's a straightforward formula, and should still be a noticable
effect.

For UI reasons, this is displayed as a single pip that goes from '.' to
'+' to '!' as you accumulate Harm.

The impact of this on the game should be incredibly minor compared to the
reams of pixels that have been spilled on the subject, but I hope it brings
joy to those it does touch.

Closes #2646.