fix: don't start resting if `can_rest_here` is false
bdabc8c6d3d1 ran afoul of one of the weird pitfalls of this code, that
rest/wait fail messaging is handled in a different place than rest/wait
start messaging. Possibly this can/should be cleaned up, but I don't
immediately know what else it would break, so I've just updated the
relevant check in _start_running to be exactly parallel.  Also commented
the point in the code where this trap can be found.