New spell: Arcjolt (L5 Conj/Air)
Fires an electrical blast that hits everything adjacent to you,
adjacent to creatures adjacent to you, etc, etc, up to the spell's
range. It also goes through metal walls, for theme reason. Damage
is exactly Fireball's, but ignoring half AC and higher variance.
(This may be too strong - my feeling is that the targeting is more
restrictive, but I could be wrong.)

No big design goal, it just seems fun.