Switch out Ice Elementalist's Hailstorm for SIB
Players have historically been a bit unhappy with Hailstorm as the
top-end of the Ice Elementalist spellset. It's an interesting spell,
but tricky to use effectively, especially without Summon Ice Beast or
Metabolic Englaciation to help maintain distance between the player
and their enemies.

That trickiness suits Reaver exactly, though, so I'm stealing the spell
from IE. In exchange, Summon Ice Beast returns to the IE start.

1b49dfe8c25f9 (2021) removed Summon Ice Beast because it was 'overly
defining' for the background (and we already have another background
defined by summons!). However, now that there's a 1-beast summon limit
on the spell and now that it's not stealing attention from Hailstorm,
that should be much less of a problem. Let's try it out!