New background: Reaver
A new warrior-mage background, Reavers primarily focus on melee combat, but
have a small set of backup spells to use for tough situations. These spells
have limitations (Drain, Contam, or simply positioning) that make them
difficult or impractical to use as a primary offense.

Reavers are intended to be a flexible background - shifting more toward melee,
casting, or neither (remaining a hybrid) as the game goes on.

Partially reverts 1cc687962fc2b6 (2011).