fix: numpad movement in local tiles
Partial revert of dbb8aaeb5fe9d6. I don't think I can entirely prevent
this from being returned in some circumstances, so I've just given up
and dealt with it in the menu code. What was happening is that two quick
events came in succession, the second suppressing the first, but the
`break` here sent processing back into a loop that needed another key to
break it.

Resolves #2644