fix: complicate to_colour_string defaults
Yet another adjustment to 29c30a7bfa8. That commit was really tailored
at dealing with passing information to webtiles as flat text. However,
to_colour_string is used internally in all sorts of places in ways that
are subtly affected by this change. Rather than try to rework those, I'm
semi-reverting 29c30a7bfa8 so that the default behavior is the same as
what it was prior to that commit, and mostly only webtiles uses of it
explicitly set a default color.