fix: remove an `mpr` overload in favor of `formatted_mpr`
These are redundant, and the `mpr` overload is used exactly once; I
think the `mpr` overload was only added (in dbf363a7764) because it's
easy to miss the existence of `formatted_mpr` (which is in a different
header). I have also prevented the implicit cast that led to the issue
in 6ad10583fd2.

The alternative, removing `formatted_mpr` and leaving only the overload,
is maybe simpler as an API, but `formatted_mpr` has a bunch more params
than regular `mpr`, and also I think it is worth someone knowing this
conversion is happening and handling it explicitly.

This commit reverts dbf363a776.