fix: handle color in "Found..." messages correctly
This had two parts. First: the use of a closing tag in message.cc is
simply incorrect, on the current state of things -- colour strings are
valid without closing tags, though closing tags must match if present,
so wrapping an open and a closing tag around an arbitrary colour string
therefore does not guarantee a valid colour string (bug introduced in
ae15d2e64f).  This at long last resolves #1224. Second: using
`formatted_mpr` in the travel.cc for some "Found..." messages is
incompatible with consistent coloring given the current state of things,
because it eliminates any sense of a default color. So even with the
first fix, messages with sub-coloring were not affected correctly by a
`message_colour` rc line (e.g. when an un-id'd scroll was in the same
message as something else).

N.b. if `formatted_string` was refactored to use a more stack-like
representation internally, as in #1249, then the second fix at least
wouldn't be necessary.  However, that PR has been languishing for a long
time, and is a bit overkill for this bug in any case, given that the use
of formatted_mpr is extremely rare (maybe should be eliminated).