fix: Correct tagging in an arrival vault (acrobat)
For kb_arrival_jail, this vault was only partially updated for species
removals and additions. It now has cells with no corresponding monster
that are also not appropriately tagged for no_tele_into and no_item_gen.
Fix this by putting in additional species monsters in the empty cells.
Some of the newer species monsters don't have appropriate tiles for use
as decorative monsters, but we'll hopefully be getting those soon.
Instead of resizing the vault only to revert that change later, fill the
empty cells with random species monsters that we can use. Also reglyph
the vault to use glyphs with no default definition.