fix: don't visit item piles with useless glowing or artefact items
On default settings, greedy explore travels to and stops at all piles
of glowing or artefact items even when the items are useless for the
character. This leads to a lot of unnecessary autoexpore interruptions
for species with gear restrictions, especially for octopodes and felids.

After this commit, greedy explore will ignore useless items if
`explore_greedy_visit` is set to `artefacts` and/or `glowing_items`
(the default). If it's set to `stacks`, explore will visit all piles,
as before.