fix: a very old off-by-one error in `join_the_dots`
I've convinced myself that this is not intentional, good to fix, and
safe to fix. We'll see -- this pathfinding method is heavily used. The
issue appears to originate from 883396a6ffc, and is that
`cur.insert(cur.begin() + random2(cur.size()), *ai);` cannot insert at
the end of `cur`. Since `cur` starts empty in the relevant loop, and at
most three things can be inserted, this insertion or the resulting
behavior is completely deterministic in a number of circumstances and at
least always deterministic in what the third element is if there are
three. The most visible one I am aware of was when pathfinding near a
masked region (a vault), if the path gets pushed N or S by a mask, and
needs to eventually get the other direction, emergent behavior can
(must?) cause the path to continue deterministically N or S until it
hits something (e.g. the map boundary), and then make a loop back, also
moving straight. This often looks quite weird, especially if there is a
long distance to the map boundary.  Example (I was able to replicate
this exact case using the seed and spent some time investigating it):

https://www.reddit.com/r/dcss/comments/wl9yrm/odd_shallow_water_formation_should_i_be_looking/

This commit also includes some useful debug code that I wouldn't want to
write again, and tries to comment the pathfinding part of this algorithm
to the best of my understanding (it's odd). I've also isolated the rng
here, to try to control the impact of calling jtd and random draws
before and after for any future changes to this algorith (since it had a
very unpredictable effect due to the varying number of draws). This may
or may not be futile.