build: use pkg-config for rltiles cross-compiling on mac
When building a universal binary or cross-compiling, this tool needs to
avoid being linked to a cross-compiled libpng dep for the other
architecture. Following the approach used for windows cross-compiling,
we just force the issue with pkg-config. This adds some extra deps to
the mac release build, but most people doing that will have them already
satisfied or easily be able to add them via homebrew or macports.