fix: send signed Xom piety to webtiles
This was maxed against 0 and then cast to an unsigned int on top of that
making it very hard to know if you are Xom's special toy on webtiles.

Followup to a1f3fc70dd8b21 which at least sent the value to webtiles,
but missed the sign issue. Resolves
https://crawl.develz.org/mantis/view.php?id=12565 and probably some
closed stuff, including #2186.