fix: make frog hop targeting exactly match fuzzing
Previously, targeting was affected weirdly by statues (see mantis
12697), because it inherited some incorrect logic from explosions. It
may have been possible to fix this simply, by changing a param in the
beam call, but I have done something a bit more ambitious: this commit
uses the exact same targeter code for both targeting and the fuzzing
itself.  Along the way it adds some more generally useful infrastructure
for doing this kind of thing, in particular, a subclass of
rectangle_iterator designed for iterating over affected squares of a
targeter.

(Also, replace a static that annoyed me in targeting code with an actual
copy constructor on FixedArray.)