fix: add a try...except block in the terminal handler
I'm seeing OSErrors on disconnect on this call on CAO in py3. I'm not
entirely sure why this isn't happening on my test device, but I'm pretty
sure it's simply that a read ends up timed badly with respect to the
crawl process ending. Also looks like CBR2 is doing this, so it's
probably a linux thing. So, treat it like an error flag.