Fixes bugs with lua skill function by using str_to_skill.
The lua skill functions were done variously, with some using
str_to_skill_safe (which returns Fighting if there is a skill
mismatch) and some using str_to_skill; and each was done
longhand.  This PR factors out the common code and uses
str_to_skill, which is the only method that makes sense.
The str_to_skill_safe gives totally false results if the
wrong skill name is used.

And as a related matter, this PR fixes the documentation,
which referred to Ranged Combat as a skill, when it is
actually called Ranged Weapons.