feat: logging options for asyncio slow callbacks
The default debug mode for logging these is fairly useless (for our code
at least, it just reports the Tornado code as the problem), and is not
ideal for production either. This commit adapts code from the aiodebug
package to do lightweight logging to help find slow callbacks.
Unfortunately I don't think it is possible to detect hotspots, or at
least doing so requires much more advanced techniques than I know about,
so to look for blocking code you need to explicitly decorate relevant
functions or code. I've done some likely candidates in this commit.

This decorator is extremely lightweight, though it's still possible I
may need to reverse this part of the commit later.