fix: make webtiles UDS socket connection fully nonblocking
On linux (or at least CAO), if the socket buffer gets filled, the server
appears to just sleep for the timeout length, and then close the socket.
Since the timeout length is 10s, this badly lags the whole server, often
causing more timeouts. I'm not sure why this is happening -- it is not
the behavior that I would expect (rather, I'd expect it to block until
the crawl binary catches up on its input), but in any case, it is bad.
There is already code that handles just dropping input if the buffer
fills, so setting this to simple non-blocking at least works on CAO in
my testing.