Responsive layout
The local tiles interface is now able to change between the normal
and small layouts as a response to a resize event.

Android devices continue using the small layout by default. Some
tablets and laptops may be big enough for the normal layout, but
more testing is required.

Other devices use the small layout only when the available size
is too small for the normal layout.