docs: separate changelog for the webtiles server
I've ended up writing out most of these in github issues at each
release, so it seems useful to reorganize this info into an actual
changelog that is stored in the repository. The main audience for this
changelog is server admins, and devs who don't normally work on the
server, so it is much more technically focused than the game changelog.
This commit goes back through my notes from these issues to fill in a
few past versions (I think basically to when I and aidanh started
actively working on the server, before this it was relatively dormant
except for bugfixes back to the original implementation). I expect I've
missed details for the old versions, though. The 0.29.0 notes are more
guaranteed to be complete.

(Also, I've taken this opportunity to impose some of my own aesthetic
changelog preferences here.)