fix: prevent massive strings in py27
I know I said that py27 is deprecated, but it's probably better not to
do it by making players lag. What was happening here is that the
annotate function gets called a lot and appends to a string, but without
the context manager in play, the string was never getting cleared,
amounting to a fairly rapid memory leak. At the rate this append
happens, this code was easily producing a string that hit memory limits,
and so I'm pretty sure once this happened each append would trigger a
non-trivial swap (and block). This commit refactors so that this
annotate call is a (nearly) noop in py27, and also adds in a truncation
sanity check (since this error case could still happen if this function
isn't called right).  Unfortunately I haven't figured out a more clever
/ elegant way to get username info into the slow callback messaging.

I've also simply removed the problem call, since it's better not to make
send_message heavy, and I believe the urgency of seeing usernames there
is removed by 5d08c370806fe (on CAO I'm currently never seeing blocking
in this function any more, which is what should be expected from now
on).