Fix bug in lua `target_weapon` -- cannot select weapon.
Crawl currently lacks the ability to speculatively examine
the to_hit value of a non-wielded weapon.  The underlying
melee_attack and ranged_attack objects appear to allow
weapon selection, but for a player, they always use the
currently wielded weapon.  This commit removes the misleading
`item` parameter and corrects the function description.