Remove Abyssal Knight
Dungeon Crawl once had many, many zealot backgrounds. The dev team
has been steadily moving away from this, because picking a god is a
big, interesting, exciting moment in each game. A zealot start has to
do something really valuable to justify taking that away.

Historically, abyssal knights' big thing was starting in the abyss. This
was a fun little minigame, a bit like delver - the bold could wander a
bit to try to pick up an extra scroll or potion, the reasonable could
race for the exit, and then you could start a normal game. It was a little
taste of something different, especially for newer players.

However, a series of changes have steadily eroded that experience. Items
no longer generate in the Abyss, and you now start on the exit. The correct
move is almost always to hit > immediately. It's a nothingburger. Sad!

Removing Abyssal Knight from the starting backgrounds allows us to make
Lugonu much stronger. Altars to Lugonu scattered around the Abyss are meant
as an incentive to worship, but the consequences of your old god's wrath
are harsh enough that it's rarely right to do so, and it also feels bad -
if you wanted to play a Lugonu worshipper, you'd just start AK! We can now
(in subsequent commits) fix both problems.

Last and certainly least, this means that adding another background to the
centre column of the background picker doesn't add a new row to the display.
:)

Closes #2689.