Remove Rolling Charge
This ability still seems quite fun, but it seems like it'll work better
as a spell (where it can combine with many different species in interesting
ways) than as a species (where it only combines with a few backgrounds in
distinctive ways). Let's move it out for now.

Also remove Armataurs' innate sinv, since that only existed for Rolling Charge
UI reasons.