New spell: Vhi's Electric Charge (L3 Tloc/Air)
As promised, the palentonga rolling charge has become a spell. Some
notable differences:

- The extra damage is electric, rather than physical. That is, enemies
  with rElec take no extra damage from an electric charge.
- Power adds a little more bonus damage, up to 33% at max power. (That is,
  a total of +133% damage at max range.)
- Since some users won't have sInv, the spell now passes through intervening
  enemies, and does its darned best to displace anything on the final space.
  (Blinking, teleporting, banishing, or just murdering as needed.)
- Stationary enemies are invalid landing targets, on the principle that they
  really shouldn't be moved and are rarely invisible. Fingers crossed.

It's in the Warper start book, replacing Gell's Gravitas, as part of an
attempt to make the start more focused on being a 'warrior-mage'. It might
be a little strong at level 3, but we can retune it as needed.

It is probably very buggy. I'm quite scared.

Tile by Sastreii.