Readjust Manifold Assault
When I first created Manifold Assault, I intended for it to be an
incentive for 'melee' characters to hybridize. In practice, it's
mostly used by transmuters, since they don't have to balance the
costs of getting a spell castable in armour. It's popular there, and
that's fine, but the original design goal isn't being accomplished.

So, lower Manifold Assault to L4 (from L5), and make it hit half as
many targets when used with UC. (It's hard to send your own limbs
through twisty space without breaking em!) We'll see how this does.

It might be cooler for this to apply contam on cast with UC instead
(per ebering). I may revise this.