fix: suppress initial mouse hover on menu display
If the mouse cursor is over a hoverable menu item when a menu pops up,
even if it isn't currently visible, a mouseenter event is generated.
This can lead to fairly unintuitive initial selections if the mouse is
in the wrong place. This commit suppresses hover entirely on initial
display. It might be nice to do this in a slightly more nuanced way, but
I think this commit is a net improvement.