Simplify tomb card
To alleviate the complexity of the card effect, I simplified the effect a little more. This card now locks the player completely at all times, regardless of power. Of course, players can still be hindered from creating walls by adjacent objects. It reduces the duration by one-third, as much as it can achieve a high shelter effect even when power is low.