feat: more butterflies
This commit makes two changes:
* Substantially increase the density of range 2 butterflies.
* Approximately double the total amount of butterflies, placing the
  second set further from the player.

Too far? Not enough? I'm not sure...

It might be more sensible to explicitly randomize based on local
density, not number, but this commit works with existing summon
placement tricks. Because it doesn't do this, this commit maybe makes
things too strong in enclosed areas? But it doesn't do anything about
butterflies dying easily.