New spell: Ray of Exothermism (L2 Conj/Fire)
Another attempt at a 'situational spell' for Reavers, Ray of Exothermism
is a twist on the old 'rod of ignition'. It fires a non-damaging ray
which causes each creature struck to ignite. But! These explosions have
a hollow centre - the creature struck isn't harmed, only those around it.

It seemed fun in playtesting - it can be quite tricky to line up a good
shot, but it's very satisfying to get one. Let's try it out.