feat: Rework slime walls
Reworks slime walls to apply 1 level of corrosion each to the player for as
long as they remain adjacent instead of doing damage; they do nothing to
monsters unless oozemancy is active, in which case they deal damage as
before.

[Committer's note: squashed commits and added handling to redraw stats.
Some further Slime changes to go alongside this would also be good - in
particular, more variation in layouts and monster pools, and possibly
adding in some more uniques. Closes #2791. -kate]