feat: let arena and rc scripts run in headless mode
The main effect of this is to lead stress tests work in headless mode.

This does absolutely no verification of whether rc files are appropriate
to run in headless mode, so a script (or just a normal rc file) that is
interactive will hang. But, an explicit rc argument most be given, so
it's on the user to verify.

Arena should be fine, it's never interactive.