proof of concept for positional fcloud
Places clouds between the player and the closest monster. Requires a better
static targeter and an ally harm/conduct check, at minimum. I originally
worked on this for a different project when fcloud was still level 6 so it
may also need balance adjustment. This significantly changes the best use
cases for the spell.