feat: allow running scripts in all builds
The script/test code is pretty entangled, so this change entails linking
some (more) debug code. But, I don't think that should be a big deal. I
can't guarantee that all scripts work (and maybe some will need to
explicitly check for the right build), but this does allow the seed
explorer to work for straightforward cases outside of debug builds.