fix: more tiles headless adjustments
Handle a bunch more ways in which code should only conditionally depend
on sdl-related state in headless mode.  This is not nearly enough to get
the full game running, and I'm not sure that'll be practical, but arena
runs, as do some rc tests. (RC test mode is still disabled in this
commit though, because it mostly crashes.)

* Default -test mode to headless for local tiles also
* Get ctrl-c working at the command line