adjust fight.cc to work when recieving cleave while wielding rift
Rift breaks the assumption that the only number besides two is one. This fixes that.

However, it still makes the assumption that the only REACH enums are just an int with a fancy title.
This will break spectacularly if a weapon is ever given a reach that isn't just a number.