feat: Adjust an oklob stair vault (CarefulOdds)
This vault is pretty polarizing among players since it guards all three
downstairs with oklobs and places in Lair. Hence it forces you to fight
through the vault to proceed, and this can involve reseting the fight
until your combat results are favorable enough. Stationary monsters are
always difficult to work with in terms of avoiding trivial resets of
fights, but if we wanted to prevent this, we'd have to remove stationary
monsters. This commit doesn't do that, but it does rebalance things a
bit so that the vault places less often, and when it does place we're
usually not forced to fight through it.

Lower the weight to 2 and have the third staircase place inside only 1/3
the time. It's not as interesting to seal off only one staircase, so
always place at least two. It's fun to torment players once and a while
with mandatory stationary monsters, but with these changes mandatory
clearing of this vault will be much more rare.

Also remove placement in Elf and Depths, since by then oklobs are a bit
less interesting as opponents, and while there's still somewhat
formidible in Elf, the theme isn't really there. Move the placement in
Snake to Swamp, where the thorn hunters this vault uses for post-Lair
make more sense, and remove the acid blobs.