fix: don't do excursions during shopping list purchases
Shop menus work with various data structures that are wiped out by a
level excursion (the shop data itself, plus item data for anything in
the shop). Buying an unid'd item could, however, trigger a level
excursion if other items of the same type are on the shopping list on
different levels. This commit refactors so that (a) there's a way of
preventing level excursions (causes a crash by default if enabled), (b)
the shopping list will detect if excursions are disallowed and store
what needs to be done, and (c) there's a way to manually trigger any
postponed excursions. This is possibly still a bit brittle if
no_excursions are used more widely, since the shopping list update will
need to be manually called in other cases. But, it's still better than
allowing arbitrary excursions as a side effect of apparently innocuous
id calls. Historically I would guess this bug was widely possible,
though I suspect it shouldn't be any more.

Resolves #2760