fix: Ds mutations conflicting with beak after 15c3b865de76
Mutations that block the same eq slot are treated as being in
physiological conflict, so after the linked commit beak also blocks
horns/antennae.  (This may not have been intentional with this commit,
it seems not to have been part of the original beak design.) However,
for whatever reason, the conflict marking for deletion needs to be
handled separately, so deletion of beak on conflict was failing. The
main consequence was that beak would block Ds horns but the resolution
code couldn't remove beak, leading to test failures (but maybe silent
failure in a real game?)

This commit adds the conflict marking to match 15c3b865de76 and prevent
crashes. However, it may be desireable to treat beak vs horns
differently, which would require some hardcoding somewhere (or
refactoring). As a side note, there's a comment here that indicates a
similar thing might have come up for antennae, but I cannot find current
code in physiology_mutation_conflict that matches this comment.

It's also possible that affected Ds games would need explicit save
compat for this fix that I haven't added (if so, they'll be missing some
horns they are due.)