feat: waypoints for all levels (gammafunk)
Two changes: relax the constraints on when waypoints can be set, and
ensure that waypoint housekeeping is correctly done for waypoints on
non-connected levels and especially in the abyss. (Luckily, I think the
conditions under which rune detection needs an update are identical to
those under which waypoints need reset.)

These are marginally useful at best for a human user on some of these
levels, so possibly only the lua API should allow e.g. waypoints in the
abyss.  But, they are useful for a bot that needs to establish reference
points.

UI note: the `G` command does not work on non-connected levels, but
waypoints can still be accessed and traveled to interactively on these
levels in a slightly clunkier way via `X`. I haven't actually checked
the lua API so if it has a restriction like this for waypoint travel,
that may be worth removing. But I have the sense that the use case of
this is really for reference points, not travel.