refactor: adjust how games are loaded
dgamelaunch-config servers currently can't load from games.d (and
because of the default values, attempting to prevents servers from
starting), but with the previous sequence of commits, the default is
never to put games configuration in the config module at all. This
commit adjusts how the double bind is resolved. By default, if `games`
is defined in the config module, games.d files are not loaded. This can
be overridden by setting the old option `use_game_yaml` to True, if
someone wants this for some reason. I've also fixed up a few issues in
how games are loaded. For backwards compatibility, the old option
`games_config_dir` is treated as a boolean-like that will have the same
effect (previously it was set to None to prevent reading games.d on
some servers). This commit also adjust the config.py documentation
accordingly and puts back in the games dictionary as an example.

(Together with previous commits) resolves #2655